Moved Shape and ShapeSorter methods to .cpp
