+++ exercises/Richard_Waite/ex02_oo_basics/src/Shape.cpp
@@ -1,0 +1,9 @@
+#include"Shape.h"
+
+void Shape::print() {
+	// std::ostream& out
+	// -> is to access members of object to which pointer points...
+	std::cout << "Type: " << this->getShapeName() << "\t" <<
+		"Area: " << this->calcArea() << "\t" <<
+		"Perimiter: " << this->calcPerimiter() << std::endl;
+}
